Update dbeaver-ce to dbeaver-ce-bin
To compile dbeaver-ce, Maven is required. Maven downloads sources from a
remote during the compile stage, which is not allowed when the
networksandbox feature is enabled in emerge. I prefer not to require
users to lower their security settings to make things compile, so
instead I opted to package up their rpm instead.